Port Self VM to 64-bit x86_64 Linux (interpreter-only, no JIT)
This is a comprehensive port of the Self VM from 32-bit i386 to native
64-bit x86_64 on Linux. The VM compiles and runs Self code correctly
in interpreter-only mode (no JIT compiler).

Key changes:

Build system:
- Add X86_64_ARCH detection in platform.cmake
- Conditionalize -m32 flags for i386-only builds
- Add runtime_stubs_x86_64.cpp with C implementations of asm stubs
- Conditionalize JIT compiler (FAST_COMPILER/SIC_COMPILER) as i386-only

Type widening (LP64):
- Widen fint/smi to int64, oopSize to 8
- Fix int32 casts of pointers throughout (use intptr_t)
- Widen parser integer types, Token fields, mark/tag bit layouts
- Fix memory subsystem (rSet, zone, space) for 64-bit address math

Snapshot loading:
- Add 32-bit-to-64-bit snapshot conversion (repack oops, expand fields)
- Handle tag bits, object sizes, and alignment for 64-bit oops

Interpreter fixes for x86_64:
- Rewrite frame model: block_scope/home_frame return 'this' (no JIT frames)
- Per-process interpreter linked list (fixes corruption during cooperative
  multitasking via SetSPAndCall)
- setjmp/longjmp NLR unwinding (replaces c_entry_point/ContinueNLRFromC)
- Fix SwitchStack ABI alignment (16-byte for SysV x86_64)

Runtime stubs (C replacements for i386 asm):
- currentFrame, FlushRegisterWindows, save1Arg
- SwitchStack0-4, SetSPAndCall
- CallPrimitiveFromInterpreter (dispatch by nargs)
- ContinueNLRFromC (fatal stub, superseded by longjmp)

+++ vm/src/i386/fast_compiler/registerState_i386.cpp
@@ -40,7 +40,7 @@ Location RegisterState::pickLocal() {
 
 void RegisterState::allocateArgs(fint nargs, bool isPrimCall) {
   Unused(isPrimCall);
-  argDepth = max(argDepth, nargs);
+  argDepth = max(argDepth, (int32)nargs);
   rcvrDepth = 1;
 }
